pyproject.toml
[project]
name = "issi-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "dvc>=3.59.1",
    "dvc-s3>=3.2.0",
    "fastapi>=0.115.12",
    "google-genai>=1.16.1",
    "polars>=1.27.1",
    "pre-commit>=4.2.0",
    "pydantic-settings>=2.9.1",
    "sentence-transformers==3.4.1",
    "spacy>=3.8.7",
    "tensorflow==2.18.0",
    "tf-keras>=2.18.0",
    "uvicorn>=0.34.2",
]
